requirements.txt
amqp==5.2.0
asgiref==3.4.1
asn1crypto==0.24.0
async-timeout==4.0.2
billiard==3.6.4.0
boto3==1.23.10
botocore==1.26.10
cached-property==1.5.2
celery==5.1.2
certifi==2018.1.18
chardet==3.0.4
click==7.1.2
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
cryptography==2.1.4
Django==3.2.19
django-debug-toolbar==3.2.4
django-filter==21.1
django-model-utils==4.2.0
django-notifications-hq==1.8.2
django-ratelimit==3.0.1
django-storages==1.12.3
djangorestframework==3.14.0
idna==2.6
importlib-metadata==4.8.3
isort==5.10.1
jmespath==0.10.0
jsonfield==3.1.0
keyring==10.6.0
keyrings.alt==3.0
kombu==5.1.0
mysqlclient==2.0.3
packaging==21.3
prompt-toolkit==3.0.36
pycrypto==2.6.1
PyGObject==3.26.1
pyparsing==3.1.1
python-dateutil==2.8.2
python-dotenv==0.20.0
python-memcached==1.59
pytz==2023.3
pyxdg==0.25
redis==4.3.6
requests==2.18.4
s3transfer==0.5.2
SecretStorage==2.3.1
six==1.11.0
sqlparse==0.4.4
ssh-import-id==5.7
swapper==1.3.0
typing_extensions==4.1.1
urllib3==1.26.16
vine==5.1.0
wcwidth==0.2.13
wrapt==1.15.0
zipp==3.6.0
